app within plasmad and updated dependenceis
